MYCE branch creation
This version of MicroGridsPy allows to model a time-evolving problem (variable demand and fuel cost along the time horizon). It also allows to split the time horizon in investment decision steps, at which the model can install additional capacity to meet the growth in demand. Results are printed in detailed and well organized Excel spreadsheets